Make JRuby and MRI share the same core ext file
