+++ lib/msgpack.rb
@@ -16,7 +16,35 @@ require "msgpack/packer"
 require "msgpack/unpacker"
 require "msgpack/factory"
 require "msgpack/symbol"
+require "msgpack/core_ext"
 
-if defined?(RUBY_ENGINE) && RUBY_ENGINE != "jruby"
-  require "msgpack/mri_ext"
+module MessagePack
+  DefaultFactory = MessagePack::Factory.new
+
+  def load(src, param = nil)
+    unpacker = nil
+
+    if src.is_a? String
+      unpacker = DefaultFactory.unpacker param
+      unpacker.feed src
+    else
+      unpacker = DefaultFactory.unpacker src, param
+    end
+
+    unpacker.full_unpack
+  end
+  alias :unpack :load
+
+  module_function :load
+  module_function :unpack
+
+  def pack(v, *rest)
+    packer = DefaultFactory.packer(*rest)
+    packer.write v
+    packer.full_pack
+  end
+  alias :dump :pack
+
+  module_function :pack
+  module_function :dump
 end
